+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.14.42
+
+* added `terminating` property to `TerminableThread`